refactor: 分离视频播放器组件，提升代码可读性
